REPO-2819 : DELETE /shared-links/{id} with admin returns 403    Revert commit 20a6ba2b281d1aa17d59008df35cb593932ed96a       Add a JUnit test into QuickShareServiceIntegrationTest
